# api/openapi.yaml
openapi: 3.1.0
info:
  title: Seva API
  description: |
    API specification for the Seva platform.
    This is the single source of truth that generates Go server stubs,
    SvelteKit TypeScript client, and Flutter Dart client.
  version: 1.0.0
  contact:
    name: Platform Engineering
    email: engineering@seva.io
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.staging.seva.io
    description: Staging
  - url: https://api.seva.io
    description: Production
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/notifications:
    get:
      operationId: listNotifications
      summary: List notifications
      description: Returns paginated notifications for the current user, unread first.
      tags: [Notifications]
      parameters:
        - name: unread_only
          in: query
          schema:
            type: boolean
            default: false
        - $ref: "#/components/parameters/PageQuery"
        - $ref: "#/components/parameters/LimitQuery"
      responses:
        "200":
          description: Notifications retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Notification"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/notifications/count:
    get:
      operationId: getUnreadNotificationCount
      summary: Get unread notification count
      tags: [Notifications]
      responses:
        "200":
          description: Unread count
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
              example:
                count: 5
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      summary: Mark notification as read
      tags: [Notifications]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Marked as read
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/notifications/read-all:
    post:
      operationId: markAllNotificationsRead
      summary: Mark all notifications as read
      tags: [Notifications]
      responses:
        "200":
          description: All marked as read
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/notifications/preferences:
    put:
      operationId: updateNotificationPreferences
      summary: Update notification preferences
      tags: [Notifications]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationPreferences"
      responses:
        "200":
          description: Preferences updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferences"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ---------------------------------------------------------------------------
  # Gamification / Points
  # ---------------------------------------------------------------------------
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from the /auth/verify-otp or /auth/refresh endpoints.

  # ---------------------------------------------------------------------------
  # Reusable Parameters
  # ---------------------------------------------------------------------------
  parameters:
...
    PageQuery:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
      example: 1
    LimitQuery:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20

  # ---------------------------------------------------------------------------
  # Reusable Responses
  # ---------------------------------------------------------------------------
  responses:
...
    Unauthorized:
      description: Unauthorized — missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: UNAUTHORIZED
            message: Authentication required
            details: null
...
    NotFound:
      description: Not found — the requested resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: NOT_FOUND
            message: Resource not found
            details: null
...
  schemas:
...
    PaginatedResponse:
      type: object
      required: [data, total, page, limit, has_more]
      properties:
        data:
          type: array
          items: {}
          description: Array of result items
        total:
          type: integer
          description: Total number of items matching the query
          examples:
            - 42
        page:
          type: integer
          description: Current page number (1-indexed)
          examples:
            - 1
        limit:
          type: integer
          description: Number of items per page
          examples:
            - 20
        has_more:
          type: boolean
          description: Whether more pages are available
          examples:
            - true
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Machine-readable error code
          examples:
            - "BAD_REQUEST"
        message:
          type: string
          description: Human-readable error message
          examples:
            - "Invalid request parameters"
        details:
          description: Additional error details (validation errors, etc.)
          oneOf:
            - type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
            - type: "null"

    # --- Notifications ---
    NotificationType:
      type: string
      enum:
        - job_new
        - job_accepted
        - job_completed
        - payment_received
        - review_received
        - dispute_update
        - points_earned
        - level_up
        - route_visit
        - seasonal_reminder
    Notification:
      type: object
      required: [id, type, title, body, created_at]
      properties:
        id:
          type: string
        type:
          $ref: "#/components/schemas/NotificationType"
        title:
          type: string
        body:
          type: string
        data:
          type: object
          description: Additional data (job_id, provider_id, etc.)
        channel:
          type: string
          enum: [push, sms, email]
        read_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    NotificationPreferences:
      type: object
      properties:
        push_enabled:
          type: boolean
          default: true
        sms_enabled:
          type: boolean
          default: true
        email_enabled:
          type: boolean
          default: false
        job_notifications:
          type: boolean
          default: true
        payment_notifications:
          type: boolean
          default: true
        marketing_notifications:
          type: boolean
          default: false

    # --- Gamification ---
...